Move benchmark utils into test_util
