+++ test_util/src/testdata/strace_summary.out
