+++ test_util/src/testdata/strace_summary2.out
